Created a Data.Semigroup.Linear module
